requirements.txt
functions-framework==3.*
google-cloud-storage==2.13.0
google-cloud-bigquery==3.*
google-ads-api==23.*
google-generativeai==0.8.*
requests==2.28.1
pandas==2.*
numpy==1.*
python-dateutil==2.*
pyyaml==6.*